+++ include/core/device/wrapper_device.hpp
