+++ include/core/utils/controls/trapezoid_profile.h
