+++ include/core/utils/math/geometry/rotation2d.h
