+++ include/core/utils/moving_average.h
