+++ include/core/vdb/crc32.hpp
