+++ include/core/vdb/registry.hpp
